Feature/product comment field

+++ src/components/product/ProductProps.tsx
@@ -40,6 +40,8 @@ interface State {
   contractTextEnglish: string;
   deliverySpecDutch: string | undefined;
   deliverySpecEnglish: string | undefined;
+  minTarget: number;
+  maxTarget: number;
 }
 
 class ProductProps extends React.Component<Props, State> {
@@ -73,6 +75,8 @@ class ProductProps extends React.Component<Props, State> {
       contractTextEnglish: product.contractTextEnglish,
       deliverySpecDutch: product.deliverySpecificationDutch,
       deliverySpecEnglish: product.deliverySpecificationEnglish,
+      minTarget: product.minTarget,
+      maxTarget: product.maxTarget,
     };
   };
 
@@ -80,7 +84,6 @@ class ProductProps extends React.Component<Props, State> {
     return new ProductParams({
       nameDutch: this.state.nameDutch,
       nameEnglish: this.state.nameEnglish,
-      targetPrice: Math.round(Number.parseFloat(this.state.targetPrice) * 100),
       status: this.state.status,
       description: this.state.description,
       categoryId: this.state.categoryId,
@@ -88,6 +91,9 @@ class ProductProps extends React.Component<Props, State> {
       contractTextEnglish: this.state.contractTextEnglish,
       deliverySpecificationEnglish: this.state.deliverySpecEnglish,
       deliverySpecificationDutch: this.state.deliverySpecDutch,
+      minTarget: this.state.minTarget,
+      maxTarget: this.state.maxTarget,
+      targetPrice: Math.round(Number.parseFloat(this.state.targetPrice) * 100),
     });
   };
 
@@ -112,7 +118,7 @@ class ProductProps extends React.Component<Props, State> {
   };
 
   remove = () => {
-    if (!this.props.create) {
+    if (!this.props.create && this.props.deleteProduct) {
       this.props.deleteProduct(this.props.product.id);
     }
   };
@@ -121,18 +127,25 @@ class ProductProps extends React.Component<Props, State> {
     if (this.props.create) {
       return undefined;
     }
-    return !(this.props.product.instances.length > 0 || this.props.product.files.length > 0);
+    return !(this.props.product.instances.length > 0
+      || this.props.product.files.length > 0);
   };
 
   render() {
     const {
       editing,
-      nameDutch, nameEnglish,
-      targetPrice, status,
+      nameDutch,
+      nameEnglish,
+      targetPrice,
+      status,
       description,
       categoryId,
-      contractTextDutch, contractTextEnglish,
-      deliverySpecDutch, deliverySpecEnglish,
+      contractTextDutch,
+      contractTextEnglish,
+      deliverySpecDutch,
+      deliverySpecEnglish,
+      minTarget,
+      maxTarget,
     } = this.state;
 
     return (
@@ -187,6 +200,24 @@ class ProductProps extends React.Component<Props, State> {
               }
             />
           </Form.Group>
+          <Form.Field
+            disabled={!editing}
+            required
+          >
+            {/* eslint-disable-next-line jsx-a11y/label-has-associated-control */}
+            <label htmlFor="form-input-category">
+              Product category
+            </label>
+            <ProductCategorySelector
+              id="form-input-category"
+              value={categoryId}
+              onChange={(val: number) => {
+                this.setState({
+                  categoryId: val,
+                });
+              }}
+            />
+          </Form.Field>
           <Form.Group widths="equal">
             <Form.Field
               disabled={!editing}
@@ -223,26 +254,47 @@ class ProductProps extends React.Component<Props, State> {
                   status:
                     data.checked ? ProductStatus.ACTIVE : ProductStatus.INACTIVE,
                 })}
-                fluid
               />
             </Form.Field>
           </Form.Group>
-          <Form.Field
-            disabled={!editing}
-            required
-          >
-            {/* eslint-disable-next-line jsx-a11y/label-has-associated-control */}
-            <label htmlFor="form-input-category">
-              Product category
-            </label>
-            <ProductCategorySelector
-              id="form-input-category"
-              value={categoryId}
-              onChange={(val: number) => this.setState({
-                categoryId: val,
-              })}
-            />
-          </Form.Field>
+          <Form.Group widths="equal">
+            <Form.Field
+              disabled={!editing}
+            >
+              {/* eslint-disable-next-line jsx-a11y/label-has-associated-control */}
+              <label htmlFor="form-input-minimal-target">
+                Minimal Target
+              </label>
+              <Input
+                id="form-input-minimal-target"
+                type="number"
+                placeholder="Minimal target"
+                value={minTarget}
+                fluid
+                onChange={(e: ChangeEvent<HTMLInputElement>) => this.setState({
+                  minTarget: e.target.value as any as number,
+                })}
+              />
+            </Form.Field>
+            <Form.Field
+              disabled={!editing}
+            >
+              {/* eslint-disable-next-line jsx-a11y/label-has-associated-control */}
+              <label htmlFor="form-input-maximum-target">
+                Maximum Target
+              </label>
+              <Input
+                id="form-input-maximum-target"
+                type="number"
+                placeholder="Maximum target"
+                value={maxTarget}
+                fluid
+                onChange={(e: ChangeEvent<HTMLInputElement>) => this.setState({
+                  maxTarget: e.target.value as any as number,
+                })}
+              />
+            </Form.Field>
+          </Form.Group>
           <Form.Field
             disabled={!editing}
           >
@@ -254,8 +306,7 @@ class ProductProps extends React.Component<Props, State> {
               id="form-input-description"
               value={description}
               onChange={(e) => this.setState({ description: e.target.value })}
-              placeholder="Comments"
-              fluid
+              placeholder="Internal comments"
             />
           </Form.Field>
           <Form.Field disabled={!editing} required error={validator.isEmpty(contractTextDutch)}>
@@ -269,8 +320,7 @@ class ProductProps extends React.Component<Props, State> {
               onChange={
                 (e) => this.setState({ contractTextDutch: e.target.value })
               }
-              placeholder="Contract Text (Dutch)"
-              fluid
+              placeholder="Contract text in Dutch"
             />
           </Form.Field>
           <Form.Field disabled={!editing} required error={validator.isEmpty(contractTextEnglish)}>
@@ -284,8 +334,7 @@ class ProductProps extends React.Component<Props, State> {
               onChange={
                 (e) => this.setState({ contractTextEnglish: e.target.value })
               }
-              placeholder="Contract Text (English)"
-              fluid
+              placeholder="Contract text in English"
             />
           </Form.Field>
           <Form.Field disabled={!editing}>
@@ -299,8 +348,7 @@ class ProductProps extends React.Component<Props, State> {
               onChange={
                 (e) => this.setState({ deliverySpecDutch: e.target.value })
               }
-              placeholder="Delivery Specification (Dutch)"
-              fluid
+              placeholder="Delivery specifications in Dutch"
             />
           </Form.Field>
           <Form.Field disabled={!editing}>
@@ -314,8 +362,7 @@ class ProductProps extends React.Component<Props, State> {
               onChange={
                 (e) => this.setState({ deliverySpecEnglish: e.target.value })
               }
-              placeholder="Delivery Specification (English)"
-              fluid
+              placeholder="Delivery specifications in English"
             />
           </Form.Field>
         </Form>
